[Refactoring] Move ExtractExpr to its own file
