+++ lib/Refactoring/CMakeLists.txt
@@ -5,12 +5,13 @@ add_swift_host_library(swiftRefactoring STATIC
   ConvertIfLetExprToGuardExpr.cpp
   ConvertToSwitchStmt.cpp
   ConvertToTernaryExpr.cpp
+  ExtractExpr.cpp
+  ExtractExprBase.cpp
   ExtractRepeatedExpr.cpp
   ExtractFunction.cpp
   MoveMembersToExtension.cpp
   Refactoring.cpp
   Renamer.cpp
-  ExtractExprBase.cpp
   ReplaceBodiesWithFatalError.cpp
   Utils.cpp
 )
