+++ lib/Refactoring/ExtractExpr.cpp
@@ -1,0 +1,42 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2014 - 2023 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+#include "ExtractExprBase.h"
+#include "RefactoringActions.h"
+
+using namespace swift::refactoring;
+
+bool RefactoringActionExtractExpr::isApplicable(const ResolvedRangeInfo &Info,
+                                                DiagnosticEngine &Diag) {
+  switch (Info.Kind) {
+  case RangeKind::SingleExpression:
+    // We disallow extract literal expression for two reasons:
+    // (1) since we print the type for extracted expression, the type of a
+    // literal may print as "int2048" where it is not typically users' choice;
+    // (2) Extracting one literal provides little value for users.
+    return checkExtractConditions(Info, Diag).success();
+  case RangeKind::PartOfExpression:
+  case RangeKind::SingleDecl:
+  case RangeKind::MultiTypeMemberDecl:
+  case RangeKind::SingleStatement:
+  case RangeKind::MultiStatement:
+  case RangeKind::Invalid:
+    return false;
+  }
+  llvm_unreachable("unhandled kind");
+}
+
+bool RefactoringActionExtractExpr::performChange() {
+  return RefactoringActionExtractExprBase(TheFile, RangeInfo, DiagEngine, false,
+                                          PreferredName, EditConsumer)
+      .performChange();
+}
